[241106] refactor: - 모달 띄우는 코드 분리 및 적용 - 분리한 useRouter 코드 적용
